Fix build.yaml and compiler warnings

+++ src/flip_bitscan.c
@@ -165,7 +165,7 @@ static const uint64_t FLIPPED_5_V[18] = {
  */
 #if __has_builtin(__builtin_subcll)
 static inline uint64_t OutflankToFlipmask(uint64_t outflank) {
-	uint64_t flipmask, cy;
+	unsigned long long flipmask, cy;
 	flipmask = __builtin_subcll(outflank, 1, 0, &cy);
 	return __builtin_addcll(flipmask, 0, cy, &cy);
 }
